@@ -2,9 +2,9 @@ name: examples_strong
 description: dart.dev strong mode examples
 
 environment:
-  sdk: '>=2.3.1 <3.0.0'
+  sdk: '>=2.10.4 <3.0.0'
 
 dev_dependencies:
   examples_util: {path: ../util}
-  pedantic: ^1.0.0
-  test: ^1.0.0
+  pedantic: ^1.9.0
+  test: ^1.14.0
