@@ -1,3 +1,3 @@
 library my_library;
 
-part "some/other/file.dart";
+part 'some/other/file.dart';
