@@ -1,8 +1,8 @@
-// ignore_for_file: sort_constructors_first, annotate_overrides
+// ignore_for_file: annotate_overrides
 // A person. The implicit interface contains greet().
 class Person {
   // In the interface, but visible only in this library.
-  final _name;
+  final String _name;
 
   // Not in the interface, since this is a constructor.
   Person(this._name);
@@ -13,7 +13,7 @@ class Person {
 
 // An implementation of the Person interface.
 class Impostor implements Person {
-  get _name => '';
+  String get _name => '';
 
   String greet(String who) => 'Hi $who. Do you know who I am?';
 }
