@@ -1,4 +1,4 @@
-// ignore_for_file: sort_constructors_first
+// ignore_for_file: prefer_initializing_formals
 import 'dart:math';
 
 class Point {
