@@ -1,4 +1,3 @@
-// ignore_for_file: sort_constructors_first
 // #docregion ''
 class Vector {
   final int x, y;
@@ -11,7 +10,8 @@ class Vector {
   // Operator == and hashCode not shown.
   // #enddocregion ''
   @override
-  bool operator ==(Object o) => o is Vector && x == o.x && y == o.y;
+  bool operator ==(Object other) =>
+      other is Vector && x == other.x && y == other.y;
 
   @override
   int get hashCode => 37 * (629 + x.hashCode) + y.hashCode;
