@@ -3,7 +3,7 @@ import 'hello.dart' deferred as hello;
 // #enddocregion import
 
 // #docregion loadLibrary
-Future greet() async {
+Future<void> greet() async {
   await hello.loadLibrary();
   hello.printGreeting();
 }
