@@ -1,4 +1,4 @@
-// ignore_for_file: unused_element, type_annotate_public_apis
+// ignore_for_file: unused_element, type_annotate_public_apis, strict_raw_type
 // #docregion import
 import 'dart:async';
 // #enddocregion import
@@ -13,7 +13,7 @@ void miscDeclAnalyzedButNotTested() {
 
   {
     // #docregion runUsingFuture
-    runUsingFuture() {
+    void runUsingFuture() {
       // ...
       findEntryPoint().then((entryPoint) {
         return runExecutable(entryPoint, args);
@@ -24,7 +24,7 @@ void miscDeclAnalyzedButNotTested() {
 
   {
     // #docregion runUsingAsyncAwait
-    runUsingAsyncAwait() async {
+    Future<void> runUsingAsyncAwait() async {
       // ...
       var entryPoint = await findEntryPoint();
       var exitCode = await runExecutable(entryPoint, args);
@@ -34,7 +34,7 @@ void miscDeclAnalyzedButNotTested() {
   }
 
   {
-    Future catchExample() async {
+    Future<void> catchExample() async {
       // #docregion catch
       var entryPoint = await findEntryPoint();
       try {
@@ -48,12 +48,12 @@ void miscDeclAnalyzedButNotTested() {
   }
 
   final url = 'humans.txt';
-  Future costlyQuery(String url) async {}
-  Future expensiveWork(dynamic value) async {}
-  Future lengthyComputation() async {}
+  Future<dynamic> costlyQuery(String url) async {}
+  Future<void> expensiveWork(dynamic value) async {}
+  Future<void> lengthyComputation() async {}
 
   {
-    Future f() {
+    Future<void> f() {
       // #docregion then-chain
       Future result = costlyQuery(url);
       result
@@ -69,7 +69,7 @@ void miscDeclAnalyzedButNotTested() {
   }
 
   {
-    Future f() async {
+    Future<void> f() async {
       // #docregion then-chain-as-await
       try {
         final value = await costlyQuery(url);
@@ -85,11 +85,11 @@ void miscDeclAnalyzedButNotTested() {
 
   bool elideBody = true;
   {
-    Future f() async {
+    Future<void> f() async {
       // #docregion wait
-      Future deleteLotsOfFiles() async => elideBody;
-      Future copyLotsOfFiles() async => elideBody;
-      Future checksumLotsOfOtherFiles() async => elideBody;
+      Future<void> deleteLotsOfFiles() async => elideBody;
+      Future<void> copyLotsOfFiles() async => elideBody;
+      Future<void> checksumLotsOfOtherFiles() async => elideBody;
 
       await Future.wait([
         deleteLotsOfFiles(),
