@@ -1,4 +1,5 @@
-// ignore_for_file: unused_element, type_annotate_public_apis
+// ignore_for_file: unused_element, type_annotate_public_apis, prefer_function_declarations_over_variables
+// ignore_for_file: avoid_function_literals_in_foreach_calls, always_declare_return_types
 import 'package:examples/language_tour/function_equality.dart'
     as function_equality;
 import 'package:test/test.dart';
@@ -61,26 +62,32 @@ void main() {
     // #enddocregion function-as-var
   });
 
-  final indexedFruit = '''0: apples
+  const indexedFruit = '''0: apples
 1: bananas
 2: oranges
 ''';
 
   test('anonymous-function', () {
-    _test() {
-      // #docregion anonymous-function
-      var list = ['apples', 'bananas', 'oranges'];
-      list.forEach((item) {
-        print('${list.indexOf(item)}: $item');
-      });
-      // #enddocregion anonymous-function
+    void _test() {
+      // #docregion anonymous-function-main
+      void main() {
+        // #docregion anonymous-function
+        const list = ['apples', 'bananas', 'oranges'];
+        list.forEach((item) {
+          print('${list.indexOf(item)}: $item');
+        });
+        // #enddocregion anonymous-function
+      }
+      // #enddocregion anonymous-function-main
+
+      main();
     }
 
     expect(_test, prints(indexedFruit));
   });
 
   test('anon-func', () {
-    _test() {
+    void _test() {
       var list = ['apples', 'bananas', 'oranges'];
       // #docregion anon-func
       list.forEach(
