@@ -4,7 +4,7 @@ import 'package:test/test.dart' as test;
 /// is an [Iterable], this matches the equivalent of `any.forEach(print)`;
 /// otherwise this matches `'$any'`.
 test.Matcher prints(dynamic any) {
-  Iterable args = any is Iterable ? any : [any];
+  final args = any is Iterable ? any : [any];
   return test.prints(args.map((arg) => '$arg').join('\n') + '\n');
 }
 
